requirements.txt
loguru
pillow
requests
markdown
python-docx
